services/air/object_detection_api/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
numpy==2.1.1
pydantic==1.10.15
python-multipart==0.0.9
ultralytics==8.3.29
opencv-python-headless
